Remove explicit dependency on window
For server-side-rendering to work we need to remove the explicit
dependency on window object because it won't be available on the server.

window-or-global package will give us the global object depending on the
current environment (window on the browser and global on node). This way
we won't get an undefined error during SSR phase.